requirements.txt
torch
torchvision
transformers
diffusers
numpy
pandas
matplotlib
tqdm
pillow
accelerate
ftfy
regex
scipy
huggingface-hub
git+https://github.com/openai/CLIP.git